requirements.txt
coloredlogs
python-dotenv
colorama
populate
fire
PyYAML